--- feature/arch/src/main/java/com/sedooj/arch/domain/repository/user/UsersNetworkRepository.kt
@@ -1,13 +1,0 @@
-package com.sedooj.architecture.domain.repository.user
-
-import com.sedooj.architecture.domain.Client
-import com.sedooj.architecture.domain.data.user.create.CreateUserInput
-import com.sedooj.architecture.domain.data.user.create.CreateUserOutput
-import com.sedooj.architecture.domain.data.user.create.auth.AuthUserInput
-
-interface UsersNetworkRepository : Client {
-
-    suspend fun createUser(input: CreateUserInput): Int
-
-    suspend fun auth(input: AuthUserInput): Int
-}
