+++ feature/arch/src/main/java/com/sedooj/arch/viewmodel/auth/SignIn.kt
@@ -3,14 +3,14 @@ package com.sedooj.architecture.viewmodel.auth
 import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.room.Room
+import com.sedooj.api.domain.data.user.create.auth.AuthUserInput
+import com.sedooj.api.domain.repository.user.UsersNetworkRepository
 import com.sedooj.arch.R
-import com.sedooj.architecture.domain.data.user.create.auth.AuthUserInput
-import com.sedooj.architecture.domain.repository.user.UsersNetworkRepository
 import com.sedooj.architecture.storage.db.AppDatabase
 import com.sedooj.architecture.storage.entity.AuthUserEntity
-import com.sedooj.architecture.viewmodel.auth.model.AuthenticationModel
-import com.sedooj.architecture.viewmodel.auth.model.AuthenticationModel.AuthState
-import com.sedooj.architecture.viewmodel.auth.model.AuthorizationInput
+import com.sedooj.arch.viewmodel.auth.model.AuthenticationModel
+import com.sedooj.arch.viewmodel.auth.model.AuthenticationModel.AuthState
+import com.sedooj.arch.viewmodel.auth.model.AuthorizationInput
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
