+++ feature/arch/src/main/java/com/sedooj/arch/viewmodel/auth/SignUp.kt
@@ -1,13 +1,13 @@
-package com.sedooj.architecture.viewmodel.auth
+package com.sedooj.arch.viewmodel.auth
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
+import com.sedooj.api.domain.data.user.create.CreateUserInput
+import com.sedooj.api.domain.repository.user.UsersNetworkRepository
 import com.sedooj.arch.R
-import com.sedooj.architecture.domain.data.user.create.CreateUserInput
-import com.sedooj.architecture.domain.repository.user.UsersNetworkRepository
-import com.sedooj.architecture.viewmodel.auth.model.AuthenticationModel
-import com.sedooj.architecture.viewmodel.auth.model.AuthenticationModel.AuthState
-import com.sedooj.architecture.viewmodel.auth.model.AuthorizationInput
+import com.sedooj.arch.viewmodel.auth.model.AuthenticationModel
+import com.sedooj.arch.viewmodel.auth.model.AuthenticationModel.AuthState
+import com.sedooj.arch.viewmodel.auth.model.AuthorizationInput
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
