+++ feature/arch/src/main/java/com/sedooj/arch/viewmodel/auth/model/AuthenticationModel.kt
@@ -1,7 +1,7 @@
-package com.sedooj.architecture.viewmodel.auth.model
+package com.sedooj.arch.viewmodel.auth.model
 
 import android.content.Context
-import com.sedooj.architecture.domain.repository.user.UsersNetworkRepository
+import com.sedooj.api.domain.repository.user.UsersNetworkRepository
 import kotlinx.coroutines.CoroutineScope
 
 interface AuthenticationModel {
