+++ feature/localstorage/build.gradle.kts
@@ -2,6 +2,7 @@ plugins {
     alias(libs.plugins.androidLibrary)
     alias(libs.plugins.kotlinAndroid)
     alias(libs.plugins.ksp)
+    id("androidx.room") version "2.6.1" apply false
 }
 
 android {
