+++ shared/api/src/main/java/com/sedooj/api/domain/data/user/create/CreateUserInput.kt
@@ -1,4 +1,4 @@
-package com.sedooj.architecture.domain.data.user.create
+package com.sedooj.api.domain.data.user.create
 
 data class CreateUserInput(
     var username: String,
